requirements.txt
asgiref==3.7.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-dotenv==1.4.2
django-keycloak==0.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
filelock==3.12.0
fonttools==4.39.4
future==0.18.3
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
oic==1.6.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.8
pyjwkest==1.4.2
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==2.16.3
python-keycloak-client==0.2.3
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.1
tzdata==2023.3
ultralytics==8.0.112
urllib3==2.0.2
zipp==3.15.0
zope.interface==6.0
